refactor: Use useModal hook in confirm modal
