+++ packages/mobile/src/hooks/useModal.ts
@@ -1,0 +1,21 @@
+import { useState } from "react";
+
+function useModal() {
+  const [ isOpen, setIsOpen ] = useState(false);
+
+  const handleModalOpen = () => {
+    return setIsOpen(true);
+  };
+
+  const handleModalClose = () => {
+    return setIsOpen(false);
+  };
+
+  return {
+    isOpen,
+    handleModalClose,
+    handleModalOpen,
+  };
+}
+
+export default useModal;
