+++ packages/mobile/src/page/Home/Restaurant/Selector/style.module.scss
@@ -51,4 +51,13 @@
       line-height: 30px;
     }
   }
+
+  .description {
+    @include typography(14);
+    display: block;
+    margin-top: 47px;
+    line-height: 30px;
+    color: $black-80;
+    text-align: center;
+  }
 }
